% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityplot.R
\name{densityplot}
\alias{densityplot}
\title{Getting a density plot during habituation and social approach}
\usage{
densityplot(df)
}
\arguments{
\item{df}{output of get_coords}
}
\value{
a plot split into a 1000 by 1000 grid of squares and then color the points by the estimated density in each square.
}
\description{
Getting a density plot during habituation and social approach
}
