% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityplot5.R
\name{densityplot5}
\alias{densityplot5}
\title{Getting a density plot in 5 min bins during habituation and social approach}
\usage{
densityplot5(df)
}
\arguments{
\item{df}{output of get_coords}
}
\value{
a 5 plot in 1 min bins split into a 1000 by 1000 grid of squares and then color the points by the estimated density in each square.
}
\description{
Getting a density plot in 5 min bins during habituation and social approach
}
