% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coords_habit.R
\name{get_coords_habit}
\alias{get_coords_habit}
\title{Get raw, plot and rescale coordinates and XY data for habituation
df  raw dataset
secs - number of seconds to keep
fr frame rate
burnin burnin period}
\usage{
get_coords_habit(df, secs = 300, fr = 30, burnin = 0)
}
\description{
Get raw, plot and rescale coordinates and XY data for habituation
df  raw dataset
secs - number of seconds to keep
fr frame rate
burnin burnin period
}
