% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Function to Get Data for X seconds from raw file.
df   raw dataframe from python output
secs number of seconds to get data for from beginning
fr  frame rate of camera
burnin - number of frames to ignore}
\usage{
get_data(df, secs = 300, fr = 30, burnin = 0)
}
\description{
Function to Get Data for X seconds from raw file.
df   raw dataframe from python output
secs number of seconds to get data for from beginning
fr  frame rate of camera
burnin - number of frames to ignore
}
