% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immobility.R
\name{immobility}
\alias{immobility}
\title{Calculate immobility}
\usage{
immobility(df, fr = 30, N = 2, X = 2)
}
\arguments{
\item{df}{output of get_coords_habit or get_coords}

\item{fr}{Frame Rate}

\item{N}{seconds that animal has not moved to count as immobile}

\item{X}{distance animal cannot have travelled for}
}
\value{
a list including i) total immobile time, ii) periods of immobility
}
\description{
Calculate immobility
}
\examples{
immobility(distcalc(get_coords_habit(habit)), X=1, N=3)
immobility(distcalc(get_coords_habit(habit)), X=2, N=1.5)
}
