% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thigmotaxis.R
\name{thigmotaxis}
\alias{thigmotaxis}
\title{thigmotaxis - center area vs non-center area. (use raw data)
use raw data in dist.5 for thigmotaxis has distance in it already}
\usage{
thigmotaxis(df, df1, xleft = 100, xright = -100, ytop = 100,
  ybottom = -100, fr = 30)
}
\description{
thigmotaxis - center area vs non-center area. (use raw data)
use raw data in dist.5 for thigmotaxis has distance in it already
}
