% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcalc.R
\name{distcalc}
\alias{distcalc}
\title{Calculate Distance Traveled}
\usage{
distcalc(df)
}
\arguments{
\item{df}{dataframe from get_coords_habit or get_coords}
}
\value{
Calculates distance from rescaleX and rescaleY values
}
\description{
Calculate Distance Traveled
}
\examples{
distcalc(df)
}
