% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coords.R
\name{get_coords}
\alias{get_coords}
\title{df  raw dataset
ROIX = distance in X of ROI
ROIY = distance in Y of ROI}
\usage{
get_coords(df, ROIX = 200, ROIY = 100)
}
\description{
df  raw dataset
ROIX = distance in X of ROI
ROIY = distance in Y of ROI
}
