% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immobility.R
\name{immobility}
\alias{immobility}
\title{Calculate immobility}
\usage{
immobility(df, fr = 30, N = 2, X = 2)
}
\description{
df output of get_coords_habit or get_coords
N seconds that animal has not moved to count as immobile
X distance animal cannot have travelled for
fr Frame Rate 30
}
