% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thigmotaxis.R
\name{thigmotaxis}
\alias{thigmotaxis}
\title{thigmotaxis}
\usage{
thigmotaxis(df, xleft = 100, xright = -100, ytop = 100,
  ybottom = -100, fr = 30)
}
\arguments{
\item{df}{dataframe from get_coords or get_coords habit}

\item{xleft}{adding x units from the left to calculate thigmotaxis}

\item{xright}{subtracting x units from the right to calculate thigmotaxis}

\item{ytop}{adding y units from the top to calculate thigmotaxis}

\item{xleft}{subtracting y units from the bottom to calculate thigmotaxis}
}
\value{
a data frame including Frame, X and Y coordinates, c_dist (center distance), t_dist (thigmotaxis distance),  c_time, t_time
}
\description{
thigmotaxis
}
\examples{
thigmotaxis(df,xleft=100, xright= -100, ytop =100, ybottom=-100, fr=30)
thigmotaxis(get_coords_habit(df),xleft=50, xright= -50, ytop =50, ybottom=-50, fr=60)
}
